requirements.txt
openpyxl==3.0.9
boto3==1.26.0
pymysql==1.0.2  # Para MySQL
google-genai==1.30.0